backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
werkzeug
requests
waitress
numpy>=1.20
Pillow>=9.0.0
mediapipe>=0.10.14
rembg>=2.0.53
onnxruntime>=1.16.0
scipy>=1.10.0
